requirements.txt
pytesseract==0.3.7
Image
pillow